Added the ResultPlotWindow and change RunningDialog to RunningWindow.
Added the ResultPlotWindow to show the result in a chart only when the
problem is of one objectives or two.
Added the RunningWindow.fxml
Delete RunningDialog and modify is controller to load a view from fxml.
Change the name of RunningDialogController a RunningWindowController.
Modified diagrama and "documento de diseño" to show the change between
RunningDialog to RunningWindow