Add parameter to PumpScheduling NSGA and Fix error
Fix error with default value of operator constructor in the
configuration windows (missing cast).
Add a new package to io utility as json.
Add a simple class to parse a json file.
Add more comment in some javadoc.
Add in FileInput the type Directory.
Add in DynamicConfigurationWindow the DirectoryChooser option.